move tests into separate file
